backend/requirements.txt
flask
flask-cors
pandas
openpyxl
google-generativeai
